Invalid commands - try using universal styles to your advantage
You've got a ton of invalid commands because you are using invalid tags. You cannot just add a number to the end of a tag like <section2> or <nav2>. This obliterates the function of that tag entirely since it isn't a real tag name - HTML and CSS will simply overlook it. You will need to remove every instance of where you do this and apply your universal styles, using Classes and IDs to target specific instances of tags. I have fixed the issue with your Nav bar and removed the times you use incorrect tags, but you will need to reformat the CSS.